Move all the primary/alice/bob helpers to lib
`wallet` and `monerod` are useful as standalone modules i.e., not for test code.
We have a bunch of helper methods that rely on first creating two accounts, one
for Alice and one for Bob - these accounts are only useful for test code.

Move all the test code helpers to `lib.rs` and export `monerod` and `wallet`
modules to the root module so that users of this crate can use them for non-test
code.
